+++ changelog.md
@@ -7,6 +7,8 @@ Last change: 25-MAY-2026 CEST
 - SPINNA: comparing models uses fitting modes and has cleaner progress dialog
 - SPINNA: convenient fitting of LE
 - Batch analysis in SPINNA for LE fitting
+- Batch analysis does not require area input (if found in metadata)
+- SPINNA: area/volume button removed (deduced automatically from densities and number of molecules in the exp. data)
 
 ## 0.10.0